Inline the mockdata in datagrid examples
