+++ projects/examples/src/components/datagrid/datagrid-filter.example.component.ts
@@ -17,7 +17,37 @@ import {
     SelectOption,
     WildCardPosition,
 } from '@vcd/ui-components';
-import { mockData, MockRecord } from './mock-data';
+
+interface MockRecord {
+    name: string;
+    city: string;
+    state: string;
+    details: {
+        gender: string;
+    };
+    age: number;
+}
+
+const mockData: MockRecord[] = [
+    {
+        name: 'Person 1',
+        city: 'Palo Alto',
+        state: 'CA',
+        details: {
+            gender: 'Male',
+        },
+        age: 30,
+    },
+    {
+        name: 'Person 2',
+        city: 'Boston',
+        state: 'MA',
+        details: {
+            gender: 'Female',
+        },
+        age: 60,
+    },
+];
 
 @Component({
     template: `
