+++ projects/examples/src/components/datagrid/datagrid-three-renderers.example.component.ts
@@ -4,15 +4,38 @@
  */
 
 import { Component } from '@angular/core';
-import {
-    BoldTextRenderer,
-    BoldTextRendererComponent,
-    ColumnComponentRendererSpec,
-    GridColumn,
-    GridDataFetchResult,
-    GridState,
-} from '@vcd/ui-components';
-import { mockData, MockRecord } from './mock-data';
+import { BoldTextRenderer, GridColumn, GridDataFetchResult, GridState } from '@vcd/ui-components';
+
+export interface MockRecord {
+    name: string;
+    city: string;
+    state: string;
+    details: {
+        gender: string;
+    };
+    age: number;
+}
+
+export const mockData: MockRecord[] = [
+    {
+        name: 'Person 1',
+        city: 'Palo Alto',
+        state: 'CA',
+        details: {
+            gender: 'Male',
+        },
+        age: 30,
+    },
+    {
+        name: 'Person 2',
+        city: 'Boston',
+        state: 'MA',
+        details: {
+            gender: 'Female',
+        },
+        age: 60,
+    },
+];
 
 @Component({
     template: `
