--- projects/examples/src/components/datagrid/mock-data.ts
@@ -1,35 +1,0 @@
-/*!
- * Copyright 2019 VMware, Inc.
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-export interface MockRecord {
-    name: string;
-    city: string;
-    state: string;
-    details: {
-        gender: string;
-    };
-    age: number;
-}
-
-export const mockData: MockRecord[] = [
-    {
-        name: 'Person 1',
-        city: 'Palo Alto',
-        state: 'CA',
-        details: {
-            gender: 'Male',
-        },
-        age: 30,
-    },
-    {
-        name: 'Person 2',
-        city: 'Boston',
-        state: 'MA',
-        details: {
-            gender: 'Female',
-        },
-        age: 60,
-    },
-];
